third_party/blink/web_tests/media/media-fragments/TC0055-expected.txt


Title: TC0055
Fragment: 't=7&t=3'
Comment: When a fragment dimensions occurs multiple times, only the last occurrence of that dimension is interpreted.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
